+++ consts.go
@@ -2,4 +2,4 @@ package qe_connector
 
 const Name = "qe-connector-go"
 
-const Version = "1.3.0"
+const Version = "1.3.1"